requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
pillow>=10.0.0
rasterio>=1.3.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
scikit-learn>=1.3.0
matplotlib>=3.7.0
pyyaml>=6.0
